Add SupportButtonContainer
