+++ src/components/PetitionSidebar/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './petition-sidebar.scss';
 import Countdown from 'components/Countdown';
-import SupportButton from 'components/SupportButton';
+import SupportButton from 'containers/SupportButton';
 import TextCenter from 'components/TextCenter';
 
 const PetitionSidebar = ({ timeMetric }) => (
