+++ src/components/SupportButton/index.js
@@ -1,9 +1,6 @@
 import React from 'react';
-import { connect } from 'react-redux';
-import { supportPetition } from 'actions/PetitionActions';
 import Button from 'components/Button';
 import settings from 'settings';
-import getPetionSupportable from 'selectors/petitionSupportable';
 
 const SupportButton = ({ petition, supportable, supportPetition }) => (
   <Button
@@ -15,21 +12,4 @@ const SupportButton = ({ petition, supportable, supportPetition }) => (
     text={settings.petition.supportButton} />
 );
 
-const mapStateToProps = ({ petition }) => ({
-  petition,
-  supportable: getPetionSupportable(petition)
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  supportPetition: (petition) => dispatch(supportPetition(petition))
-});
-
-SupportButton.propTypes = {
-  petition: React.PropTypes.object,
-  supportable: React.PropTypes.bool
-};
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(SupportButton);
+export default SupportButton;
